[vis|k3d] refactor mini-lib back into pymor namespace
